part_1-python-fundamentals/01_mini-projects/12_hello-azure-storage-acct/requirements.txt
azure-cosmosdb-table==1.0.6
azurerm==0.10.0
azure-storage-queue==12.7.3
python-dotenv==1.0.0
